got really spun with all these different examples; trying to work with redux docs example to see if refactoring it will help me understand better
