+++ todo/src/App.js
@@ -1,7 +1,26 @@
-import React from 'react'
+import React from 'react';
+import TodoList from './TodoList';
+import AddTodo from '../containers/AddTodo';
 
+​
 const App = () => (
-    <div>
-        <
-    </div>
+  <div>
+    <AddTodo />
+    <TodoList />
+   
+  </div>
 )
+​
+export default App
+
+
+// import React from 'react'
+// import ToDoList from './components/TodoList'
+
+// const App = () => (
+//     <div>
+//         <ToDoList />
+//     </div>
+// )
+
+// export default App;
